+++ pkg/api/convertertoapi.go
@@ -7,6 +7,7 @@ import (
 	"github.com/Azure/acs-engine/pkg/api/v20160930"
 	"github.com/Azure/acs-engine/pkg/api/v20170131"
 	"github.com/Azure/acs-engine/pkg/api/v20170701"
+	"github.com/Azure/acs-engine/pkg/api/v20170930"
 	"github.com/Azure/acs-engine/pkg/api/vlabs"
 )
 
@@ -121,6 +122,31 @@ func ConvertVLabsContainerService(vlabs *vlabs.ContainerService) *ContainerServi
 	c.Properties = &Properties{}
 	convertVLabsProperties(vlabs.Properties, c.Properties)
 	return c
+}
+
+// ConvertV20170930UpgradeContainerService converts a v20170930 UpgradeContainerService to an unversioned UpgradeContainerService
+func ConvertV20170930UpgradeContainerService(vUCS *v20170930.UpgradeContainerService) *UpgradeContainerService {
+	ucs := &UpgradeContainerService{}
+	switch vUCS.OrchestratorType {
+	case v20170930.Kubernetes:
+		ucs.OrchestratorType = Kubernetes
+	case v20170930.DCOS:
+		ucs.OrchestratorType = DCOS
+	case v20170930.Swarm:
+		ucs.OrchestratorType = Swarm
+	case v20170930.DockerCE:
+		ucs.OrchestratorType = SwarmMode
+	}
+	ucs.OrchestratorRelease = vUCS.OrchestratorRelease
+	ucs.OrchestratorVersion = vUCS.OrchestratorVersion
+	return ucs
+}
+
+// ConvertVLabsUpgradeContainerService converts a vlabs UpgradeContainerService to an unversioned UpgradeContainerService
+func ConvertVLabsUpgradeContainerService(vlabUCS *vlabs.UpgradeContainerService) *UpgradeContainerService {
+	ucs := &UpgradeContainerService{}
+	convertVLabsOrchestratorProfile((*vlabs.OrchestratorProfile)(vlabUCS), (*OrchestratorProfile)(ucs))
+	return ucs
 }
 
 // convertV20160930ResourcePurchasePlan converts a v20160930 ResourcePurchasePlan to an unversioned ResourcePurchasePlan
